requirements.txt
langchain-core>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-community>=0.3,<0.4
python-dotenv>=1.0,<2.0

# ✅ Otimizado para AWS - SEM embeddings caros!
# Removido:
#  - langchain-text-splitters (usado com FAISS)
#  - langchain-community[vectorstores] (FAISS, Pinecone, etc)
#
# Porque:
# - BuscaTextoLocal usa busca simples (sem custo)
# - Cache em memória economiza chamadas ao LLM
# - Histórico resumido = menos tokens
faiss-cpu>=1.7,<2.0
sqlalchemy>=2.0,<3.0

# 🧪 Testes e QA
pytest>=7.0,<8.0
pytest-asyncio>=0.21,<0.22
pytest-cov>=4.0,<5.0
pytest-mock>=3.10,<3.11
responses>=0.23,<0.24